Arreglos en manager servicio

+++ src/main/java/com/proyectoDualWebService/dto/Servicio.java
@@ -14,7 +14,7 @@ public class Servicio {
     private int id;
     private String nombre;
     private String plan;
-    private float precio;
+    private double precio;
     private String web;
 
 
@@ -23,7 +23,7 @@ public class Servicio {
             this.id = result.getInt("id");
             this.nombre = result.getString("nombre");
             this.plan = result.getString("plan");
-            this.precio = result.getFloat("precio");
+            this.precio = result.getDouble("precio");
             this.web = result.getString("web");
 
         } catch (SQLException e) {
